TextualSummarizationOfVideosInIndicLanguage-backend/requirements.txt
flask
flask-cors
gtts
langdetect
requests
transformers
yt-dlp
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
